refactor(orchestra): split dispatch_queue_helpers.py by responsibility
Split dispatch_queue_helpers.py (325 lines, 8 functions, 5 responsibilities)
into focused modules to improve code organization and maintainability.

Changes:
- NEW orchestra/issue_loader.py (80 LOC): issue loading and flow context
  - find_role_for_state, is_auto_task_branch, get_flow_context, load_issue
- NEW orchestra/queue_operations.py (190 LOC): queue operations
  - select_ready_issues, promote_progressed_entries
- MODIFY utils/label_utils.py (+30 LOC): added clean_old_state_labels
- DELETE orchestra/dispatch_queue_helpers.py: responsibilities fully decomposed
- DEDUPLICATE: removed duplicate normalize_labels, unified to label_utils.py version

This is a pure refactoring with no behavior changes. All 157 tests pass.

Issue: #772
